+++ src/main/java/com/bandeira/clinica_alves_oliveira/models/Procedure.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,13 +25,14 @@ public class Procedure {
 
     private Specialty specialty;
 
-    private Double value;
+    private BigDecimal value;
 
 
     @OneToMany(mappedBy = "procedure")
     List<Budget> budgets = new ArrayList<>();
 
-    public Procedure(String description, Specialty specialty, Double value) {
+
+    public Procedure(String description, Specialty specialty, BigDecimal value) {
         this.description = description;
         this.specialty = specialty;
         this.value = value;
